# M2-Challenge-Rivera-Nelson/RestAPI/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /subtract:
    post:
      tags:
        - math-solution-controller
      operationId: subtraction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
  /multiply:
    post:
      tags:
        - math-solution-controller
      operationId: multiplication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
  /divide:
    post:
      tags:
        - math-solution-controller
      operationId: division
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
  /add:
    post:
      tags:
        - math-solution-controller
      operationId: addition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
  /randomMonth:
    get:
      tags:
        - month-controller
      operationId: randomMonth
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
  /month/{monthNumber}:
    get:
      tags:
        - month-controller
      operationId: randomMonth_1
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    MathSolution:
      type: object
      properties:
        operand1:
          type: integer
          format: int32
        operand2:
          type: integer
          format: int32
        operation:
          type: string
        answer:
          type: integer
          format: int32
    Month:
      type: object
      properties:
        number:
          type: integer
          format: int32
        name:
          type: string